[FEAT]좋아요 클릭 오류해결!

+++ app/src/main/java/com/example/carrotmarket/data/Product.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Product(
-    val id : Long,              // id
+    val id : Int,              // id
     @DrawableRes
     val image : Int,            // 이미지
     val name : String,          // 제품먕
